Changed default port. Make debug mode a CLI flag
CLI flag means we can invoke the base webapp with the '--debug'
flag to put the Flask server into debug mode. This means any
changes to code in the Flask app will trigger a reload of the
running server, without that server stopping. This is handy.

The default port is now in the range that doesn't require root
privileges.